I continue to fucking up with my xml parser

+++ database/migrations/2019_04_24_000000_create_collection_table.php
@@ -14,8 +14,8 @@ class CreateCollectionTable extends Migration
 	public function up()
 	{
 		Schema::create('collection', function (Blueprint $table) {
-			$table->increments('id');
-			$table->integer('parent_id')->unsigned()->nullable();
+			$table->bigIncrements('id');
+			$table->bigInteger('parent_id')->unsigned()->nullable();
 			$table->integer('updated_at')->unsigned();
 			$table->integer('created_at')->unsigned();
